requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# archivy/__init__.py: 7
# archivy/check_changes.py: 5
# archivy/data.py: 9
# archivy/extensions.py: 6
# archivy/models.py: 13
# archivy/routes.py: 7
Flask == 1.1.2

# archivy/forms.py: 2
Flask_WTF == 0.14.3

# archivy/forms.py: 4
WTForms == 2.3.1

# archivy/data.py: 10
Werkzeug == 1.0.1

# archivy/config.py: 2
appdirs == 1.4.4

# archivy/models.py: 10
attrs == 20.2.0

# archivy/models.py: 11
beautifulsoup4 >= 4.8.2


# archivy/__init__.py: 6
# archivy/extensions.py: 4,5
elasticsearch == 7.7.1


# archivy/__init__.py: 8
# archivy/check_changes.py: 6
# archivy/models.py: 12
# archivy/routes.py: 5
pypandoc == 1.5

# archivy/run.py: 1
python_dotenv == 0.13.0

# archivy/data.py: 8
# archivy/models.py: 5

python_frontmatter == 0.5.0

# archivy/models.py: 7
# archivy/routes.py: 4
requests == 2.24.0

# archivy/extensions.py: 7
# archivy/routes.py: 5
tinydb == 4.1.1

# archivy/models.py: 8
validators == 0.15.0

# archivy/check_changes.py: 7,8
watchdog == 0.10.3

# archivy/__init__.py: 9
# archivy/models.py: 13
# archivy/routes.py: 10
flask-login == 0.5.0

click_plugins

